api cleanup, adding cancelation toekn, refactor of tests
